refactor: Move utils.SupportedCI to cmd.secret.scan.ci
This is the only place it is used.
